api/requirements.txt
alembic==1.12.1
Flask==3.0.0
Flask_Cors==4.0.0
flask_marshmallow==0.15.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
marshmallow==3.20.1
psycopg2_binary==2.9.9
PyJWT==2.8.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
